Fix cross reference
