+++ tests/gnmi/conftest.py
@@ -12,7 +12,8 @@ from tests.common.helpers.dut_utils import check_container_state
 from tests.gnmi.helper import gnmi_container, apply_cert_config, recover_cert_config, create_ext_conf, create_ca_conf
 from tests.gnmi.helper import GNMI_SERVER_START_WAIT_TIME, check_ntp_sync_status
 from tests.common.gu_utils import create_checkpoint, rollback
-from tests.common.helpers.gnmi_utils import GNMIEnvironment
+from tests.common.helpers.gnmi_utils import GNMIEnvironment, create_revoked_cert_and_crl, \
+                                            create_gnmi_certs, delete_gnmi_certs
 from tests.common.helpers.ntp_helper import setup_ntp_context
 
 
@@ -59,175 +60,6 @@ def setup_gnmi_ntp_client_server(duthosts, rand_one_dut_hostname, ptfhost):
 
     with setup_ntp_context(ptfhost, duthost, False):
         yield
-
-
-def create_revoked_cert_and_crl(localhost, ptfhost):
-    # Create client key
-    local_command = "openssl genrsa -out gnmiclient.revoked.key 2048"
-    localhost.shell(local_command)
-
-    # Create client CSR
-    local_command = "openssl req \
-                        -new \
-                        -key gnmiclient.revoked.key \
-                        -subj '/CN=test.client.revoked.gnmi.sonic' \
-                        -out gnmiclient.revoked.csr"
-    localhost.shell(local_command)
-
-    # Sign client certificate
-    crl_url = "http://{}:1234/crl".format(ptfhost.mgmt_ip)
-    create_ca_conf(crl_url, "crlext.cnf")
-    local_command = "openssl x509 \
-                        -req \
-                        -in gnmiclient.revoked.csr \
-                        -CA gnmiCA.pem \
-                        -CAkey gnmiCA.key \
-                        -CAcreateserial \
-                        -out gnmiclient.revoked.crt \
-                        -days 825 \
-                        -sha256 \
-                        -extensions req_ext -extfile crlext.cnf"
-    localhost.shell(local_command)
-
-    # create crl config file
-    local_command = "rm -f gnmi/crl/index.txt"
-    localhost.shell(local_command)
-    local_command = "touch gnmi/crl/index.txt"
-    localhost.shell(local_command)
-
-    local_command = "rm -f gnmi/crl/sonic_crl_number"
-    localhost.shell(local_command)
-    local_command = "echo 00 > gnmi/crl/sonic_crl_number"
-    localhost.shell(local_command)
-
-    # revoke cert CRL
-    local_command = "openssl ca \
-                        -revoke gnmiclient.revoked.crt \
-                        -keyfile gnmiCA.key \
-                        -cert gnmiCA.pem \
-                        -config gnmi/crl/crl.cnf"
-
-    localhost.shell(local_command)
-
-    # re-create CRL
-    local_command = "openssl ca \
-                        -gencrl \
-                        -keyfile gnmiCA.key \
-                        -cert gnmiCA.pem \
-                        -out sonic.crl.pem \
-                        -config gnmi/crl/crl.cnf"
-
-    localhost.shell(local_command)
-
-    # copy to PTF for test
-    ptfhost.copy(src='gnmiclient.revoked.crt', dest='/root/')
-    ptfhost.copy(src='gnmiclient.revoked.key', dest='/root/')
-    ptfhost.copy(src='sonic.crl.pem', dest='/root/')
-    ptfhost.copy(src='gnmi/crl/crl_server.py', dest='/root/')
-
-    local_command = "rm \
-                        crlext.cnf \
-                        gnmi/crl/index.* \
-                        gnmi/crl/sonic_crl_number.*"
-    localhost.shell(local_command)
-
-
-def create_gnmi_certs(duthost, localhost, ptfhost):
-    '''
-    Create GNMI client certificates
-    '''
-    # Create Root key
-    local_command = "openssl genrsa -out gnmiCA.key 2048"
-    localhost.shell(local_command)
-
-    # Create Root cert
-    local_command = "openssl req \
-                        -x509 \
-                        -new \
-                        -nodes \
-                        -key gnmiCA.key \
-                        -sha256 \
-                        -days 1825 \
-                        -subj '/CN=test.gnmi.sonic' \
-                        -out gnmiCA.pem"
-    localhost.shell(local_command)
-
-    # Create server key
-    local_command = "openssl genrsa -out gnmiserver.key 2048"
-    localhost.shell(local_command)
-
-    # Create server CSR
-    local_command = "openssl req \
-                        -new \
-                        -key gnmiserver.key \
-                        -subj '/CN=test.server.gnmi.sonic' \
-                        -out gnmiserver.csr"
-    localhost.shell(local_command)
-
-    # Sign server certificate
-    create_ext_conf(duthost.mgmt_ip, "extfile.cnf")
-    local_command = "openssl x509 \
-                        -req \
-                        -in gnmiserver.csr \
-                        -CA gnmiCA.pem \
-                        -CAkey gnmiCA.key \
-                        -CAcreateserial \
-                        -out gnmiserver.crt \
-                        -days 825 \
-                        -sha256 \
-                        -extensions req_ext -extfile extfile.cnf"
-    localhost.shell(local_command)
-
-    # Create client key
-    local_command = "openssl genrsa -out gnmiclient.key 2048"
-    localhost.shell(local_command)
-
-    # Create client CSR
-    local_command = "openssl req \
-                        -new \
-                        -key gnmiclient.key \
-                        -subj '/CN=test.client.gnmi.sonic' \
-                        -out gnmiclient.csr"
-    localhost.shell(local_command)
-
-    # Sign client certificate
-    local_command = "openssl x509 \
-                        -req \
-                        -in gnmiclient.csr \
-                        -CA gnmiCA.pem \
-                        -CAkey gnmiCA.key \
-                        -CAcreateserial \
-                        -out gnmiclient.crt \
-                        -days 825 \
-                        -sha256"
-    localhost.shell(local_command)
-
-    create_revoked_cert_and_crl(localhost, ptfhost)
-
-    # Copy CA certificate, server certificate and client certificate over to the DUT
-    duthost.copy(src='gnmiCA.pem', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiserver.crt', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiserver.key', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiclient.crt', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiclient.key', dest='/etc/sonic/telemetry/')
-    # Copy CA certificate and client certificate over to the PTF
-    ptfhost.copy(src='gnmiCA.pem', dest='/root/')
-    ptfhost.copy(src='gnmiclient.crt', dest='/root/')
-    ptfhost.copy(src='gnmiclient.key', dest='/root/')
-
-    create_checkpoint(duthost, SETUP_ENV_CP)
-
-
-def delete_gnmi_certs(localhost):
-    '''
-    Delete GNMI client certificates
-    '''
-    local_command = "rm \
-                        extfile.cnf \
-                        gnmiCA.* \
-                        gnmiserver.* \
-                        gnmiclient.*"
-    localhost.shell(local_command)
 
 
 @pytest.fixture(scope="module", autouse=True)
