+++ tests/gnmi/test_gnmi_configdb.py
@@ -8,7 +8,7 @@ import time
 from .helper import gnmi_set, gnmi_get, gnoi_reboot
 from .helper import gnmi_subscribe_polling
 from .helper import gnmi_subscribe_streaming_sample, gnmi_subscribe_streaming_onchange
-from .helper import add_gnmi_client_common_name
+from tests.common.helpers.gnmi_utils import add_gnmi_client_common_name
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.utilities import wait_until
 from tests.common.platform.processes_utils import wait_critical_processes
